Moved isAdjacent method to Position, fixed all calls to this method and tests too. < organisation
